fix #69: move code into src/ for web-ext build
