+++ src/js/canonicalize.js
