+++ src/js/disconnect-blocklist.json
