+++ src/manifest.json
@@ -1,7 +1,7 @@
 {
   "manifest_version": 2,
   "name": "Tracking Protection Experiment",
-  "version": "1.0rc1",
+  "version": "1.0rc2",
 
   "description": "Implements tracking protection as a web extension add-on.",
   "icons": {
